release-assistant/requirement.txt
pyyaml
requests
unittest